@@ -78,25 +78,27 @@
     ]
   },
   "devDependencies": {
-    "@storybook/addon-essentials": "7.0.0-rc.1",
-    "@storybook/addon-interactions": "7.0.0-rc.1",
-    "@storybook/addon-links": "7.0.0-rc.1",
-    "@storybook/blocks": "7.0.0-rc.1",
-    "@storybook/react": "7.0.0-rc.1",
-    "@storybook/react-vite": "7.0.0-rc.1",
-    "@storybook/testing-library": "^0.0.14--canary.30.5f44d14.0",
-    "@vitejs/plugin-react-swc": "^3.2.0",
+    "@storybook/addon-essentials": "7.0.17",
+    "@storybook/addon-interactions": "7.0.17",
+    "@storybook/addon-links": "7.0.17",
+    "@storybook/blocks": "7.0.17",
+    "@storybook/react": "7.0.17",
+    "@storybook/react-vite": "7.0.17",
+    "@storybook/testing-library": "0.1.0",
+    "@vitejs/plugin-react-swc": "^3.3.1",
     "autoprefixer": "^10.4.14",
-    "jest": "^29.4.3",
+    "jest": "^29.5.0",
     "postcss": "^8.4.16",
-    "prettier": "^2.8.4",
-    "prettier-plugin-tailwindcss": "^0.2.4",
+    "prettier": "^2.8.8",
+    "prettier-plugin-tailwindcss": "^0.3.0",
     "source-map-explorer": "^2.5.3",
-    "storybook": "7.0.0-rc.1",
-    "tailwindcss": "^3.2.7",
-    "ts-jest": "^29.0.5",
-    "typescript": "^4.9.5",
-    "vite": "^4.1.4",
-    "vite-plugin-compression": "^0.5.1"
+    "storybook": "7.0.17",
+    "tailwindcss": "^3.3.2",
+    "ts-jest": "^29.1.0",
+    "typescript": "^5.0.4",
+    "vite": "^4.2.2",
+    "vite-imagetools": "^5.0.4",
+    "vite-plugin-compression": "^0.5.1",
+    "vite-plugin-rewrite-all": "^1.0.1"
   }
 }
