@@ -1,5 +1,6 @@
 import React from "react";
-import Otter from "./otter.png";
+// @ts-expect-error
+import Otter from "./otter.png?w=128&h=128&webp";
 
 const Logo: React.FC = () => (
   <div className="flex cursor-default items-center justify-center space-x-4 font-title text-6xl font-bold text-link-blue">
