@@ -1,11 +1,12 @@
 import React from "react";
 import ContentFrame from "../../components/ContentFrame";
 import StandardSelectionBoundary from "../../selection/StandardSelectionBoundary";
-import PageControl from "../../search/PageControl";
+import SearchResultNavBar from "../../search/SearchResultNavBar";
 import ResultHeader from "../../search/ResultHeader";
 import PendingResults from "../../search/PendingResults";
 import TransactionItem from "../../search/TransactionItem";
 import { useFeeToggler } from "../../search/useFeeToggler";
+import { totalTransactionsFormatter } from "../../search/messages";
 import { ProcessedTransaction } from "../../types";
 import { PAGE_SIZE } from "../../params";
 
@@ -24,20 +25,12 @@ const BlockTransactionResults: React.FC<BlockTransactionResultsProps> = ({
 
   return (
     <ContentFrame>
-      <div className="flex items-baseline justify-between py-3">
-        <div className="text-sm text-gray-500">
-          {page === undefined ? (
-            <>Waiting for search results...</>
-          ) : (
-            <>A total of {total} transactions found</>
-          )}
-        </div>
-        <PageControl
-          pageNumber={pageNumber}
-          pageSize={PAGE_SIZE}
-          total={total}
-        />
-      </div>
+      <SearchResultNavBar
+        pageNumber={pageNumber}
+        pageSize={PAGE_SIZE}
+        total={total}
+        totalFormatter={totalTransactionsFormatter}
+      />
       <ResultHeader
         feeDisplay={feeDisplay}
         feeDisplayToggler={feeDisplayToggler}
@@ -47,16 +40,12 @@ const BlockTransactionResults: React.FC<BlockTransactionResultsProps> = ({
           {page.map((tx) => (
             <TransactionItem key={tx.hash} tx={tx} feeDisplay={feeDisplay} />
           ))}
-          <div className="flex items-baseline justify-between py-3">
-            <div className="text-sm text-gray-500">
-              A total of {total} transactions found
-            </div>
-            <PageControl
-              pageNumber={pageNumber}
-              pageSize={PAGE_SIZE}
-              total={total}
-            />
-          </div>
+          <SearchResultNavBar
+            pageNumber={pageNumber}
+            pageSize={PAGE_SIZE}
+            total={total}
+            totalFormatter={totalTransactionsFormatter}
+          />
         </StandardSelectionBoundary>
       ) : (
         <PendingResults />
