@@ -1,95 +1,31 @@
-import { useContext, FC } from "react";
-import { useSearchParams } from "react-router-dom";
-import { commify } from "@ethersproject/units";
-import StandardFrame from "../components/StandardFrame";
-import StandardSubtitle from "../components/StandardSubtitle";
-import ContentFrame from "../components/ContentFrame";
-import StandardSelectionBoundary from "../selection/StandardSelectionBoundary";
-import StandardTable from "../components/StandardTable";
-import StandardTHead from "../components/StandardTHead";
-import StandardTBody from "../components/StandardTBody";
-import PageControl from "../search/PageControl";
-import ERC721Item from "./ERC721Item";
-import { RuntimeContext } from "../useRuntime";
-import { useERC721Count, useERC721List } from "../useErigonHooks";
+import { FC } from "react";
+import GenericContractSearchResult from "./GenericContractSearchResult";
+import ERC721Header from "./ERC721Header";
+import ERC721Item, { mapper } from "./ERC721Item";
+import { useContractSearchPage } from "../ots2/useUIHooks";
+import { erc721MatchParser } from "../ots2/contractMatchParsers";
 import { PAGE_SIZE } from "../params";
 
 const AllERC721: FC = () => {
-  const { provider } = useContext(RuntimeContext);
-
-  const [searchParams] = useSearchParams();
-  let pageNumber = 1;
-  const p = searchParams.get("p");
-  if (p) {
-    try {
-      pageNumber = parseInt(p);
-    } catch (err) {}
-  }
-
-  const total = useERC721Count(provider);
-  const page = useERC721List(provider, pageNumber, PAGE_SIZE);
+  const { pageNumber, page, total } = useContractSearchPage(
+    "ERC721",
+    erc721MatchParser,
+    mapper
+  );
 
   document.title = `ERC721 Tokens | Otterscan`;
 
   return (
-    <StandardFrame>
-      <StandardSubtitle>
-        <div className="flex items-baseline space-x-1">
-          <span>ERC721 tokens</span>
-        </div>
-      </StandardSubtitle>
-      <ContentFrame key={pageNumber}>
-        <div className="flex items-baseline justify-between py-3">
-          <div className="text-sm text-gray-500">
-            {page === undefined || total === undefined ? (
-              <>Waiting for search results...</>
-            ) : (
-              <>A total of {commify(total)} contracts found</>
-            )}
-          </div>
-          {total !== undefined && (
-            <PageControl
-              pageNumber={pageNumber}
-              pageSize={PAGE_SIZE}
-              total={total}
-            />
-          )}
-        </div>
-        <StandardTable>
-          <StandardTHead>
-            <th className="w-96">Address</th>
-            <th className="w-28">Block</th>
-            <th className="w-40">Age</th>
-            <th>Name</th>
-            <th>Symbol</th>
-          </StandardTHead>
-          {page !== undefined ? (
-            <StandardSelectionBoundary>
-              <StandardTBody>
-                {page.map((m) => (
-                  <ERC721Item key={m.address} m={m} />
-                ))}
-              </StandardTBody>
-            </StandardSelectionBoundary>
-          ) : (
-            // <PendingResults />
-            <></>
-          )}
-        </StandardTable>
-        {page !== undefined && total !== undefined && (
-          <div className="flex items-baseline justify-between py-3">
-            <div className="text-sm text-gray-500">
-              A total of {commify(total)} contracts found
-            </div>
-            <PageControl
-              pageNumber={pageNumber}
-              pageSize={PAGE_SIZE}
-              total={total}
-            />
-          </div>
-        )}
-      </ContentFrame>
-    </StandardFrame>
+    <GenericContractSearchResult
+      title="ERC721 tokens"
+      header={<ERC721Header />}
+      cols={5}
+      pageNumber={pageNumber}
+      pageSize={PAGE_SIZE}
+      total={total}
+      page={page}
+      Item={(m) => <ERC721Item {...m} />}
+    />
   );
 };
 
