@@ -1,31 +1,38 @@
-import { FC, memo, useContext } from "react";
+import { FC, memo } from "react";
 import DecoratedAddressLink from "../execution/components/DecoratedAddressLink";
 import BlockLink from "../components/BlockLink";
 import TimestampAge from "../components/TimestampAge";
-import { RuntimeContext } from "../useRuntime";
-import { ContractMatch, useBlockData } from "../useErigonHooks";
+import { ChecksummedAddress } from "../types";
+import { ResultMapper } from "../ots2/useUIHooks";
 
 type ContractItemProps = {
-  m: ContractMatch;
+  blockNumber: number;
+  timestamp: number;
+  address: ChecksummedAddress;
 };
+
+export const mapper: ResultMapper<ContractItemProps> = (m, blocksSummary) => ({
+  blockNumber: m.blockNumber,
+  timestamp: blocksSummary.get(m.blockNumber)!.timestamp,
+  address: m.address,
+});
 
 const ContractItem: FC<ContractItemProps> = ({
-  m: { address, blockNumber },
-}) => {
-  const { provider } = useContext(RuntimeContext);
-  const block = useBlockData(provider, blockNumber.toString());
-
-  return (
-    <tr>
-      <td>
-        <DecoratedAddressLink address={address} plain />
-      </td>
-      <td>
-        <BlockLink blockTag={blockNumber} />
-      </td>
-      <td>{block && <TimestampAge timestamp={block.timestamp} />}</td>
-    </tr>
-  );
-};
+  blockNumber,
+  timestamp,
+  address,
+}) => (
+  <>
+    <td>
+      <DecoratedAddressLink address={address} plain />
+    </td>
+    <td>
+      <BlockLink blockTag={blockNumber} />
+    </td>
+    <td>
+      <TimestampAge timestamp={timestamp} />
+    </td>
+  </>
+);
 
 export default memo(ContractItem);
