@@ -1,52 +1,55 @@
-import { FC, memo, useContext } from "react";
+import { FC, memo } from "react";
 import DecoratedAddressLink from "../execution/components/DecoratedAddressLink";
 import BlockLink from "../components/BlockLink";
 import TimestampAge from "../components/TimestampAge";
 import CanBeEmptyText from "../components/CanBeEmptyText";
-import { RuntimeContext } from "../useRuntime";
-import {
-  ContractMatch,
-  useBlockData,
-  useERC20Metadata,
-} from "../useErigonHooks";
+import { ChecksummedAddress } from "../types";
+import { ResultMapper } from "../ots2/useUIHooks";
 
-type ERC20temProps = {
-  m: ContractMatch;
+type ERC20ItemProps = {
+  blockNumber: number;
+  timestamp: number;
+  address: ChecksummedAddress;
+  name: string;
+  symbol: string;
+  decimals: number;
 };
 
-const ERC20Item: FC<ERC20temProps> = ({ m: { address, blockNumber } }) => {
-  const { provider } = useContext(RuntimeContext);
-  const erc20meta = useERC20Metadata(provider, address, blockNumber);
-  const block = useBlockData(provider, blockNumber.toString());
+export const mapper: ResultMapper<ERC20ItemProps> = (m, blocksSummary) => ({
+  blockNumber: m.blockNumber,
+  timestamp: blocksSummary.get(m.blockNumber)!.timestamp,
+  address: m.address,
+  name: m.name,
+  symbol: m.symbol,
+  decimals: m.decimals,
+});
 
-  return (
-    <tr>
-      <td>
-        <DecoratedAddressLink address={address} plain />
-      </td>
-      <td>
-        <BlockLink blockTag={blockNumber} />
-      </td>
-      <td>{block && <TimestampAge timestamp={block.timestamp} />}</td>
-      {erc20meta ? (
-        <>
-          <td>
-            <CanBeEmptyText text={erc20meta.name} />
-          </td>
-          <td>
-            <CanBeEmptyText text={erc20meta.symbol} />
-          </td>
-          <td>{erc20meta.decimals}</td>
-        </>
-      ) : (
-        <>
-          <td></td>
-          <td></td>
-          <td></td>
-        </>
-      )}
-    </tr>
-  );
-};
+const ERC20Item: FC<ERC20ItemProps> = ({
+  blockNumber,
+  timestamp,
+  address,
+  name,
+  symbol,
+  decimals,
+}) => (
+  <>
+    <td>
+      <DecoratedAddressLink address={address} plain />
+    </td>
+    <td>
+      <BlockLink blockTag={blockNumber} />
+    </td>
+    <td>
+      <TimestampAge timestamp={timestamp} />
+    </td>
+    <td>
+      <CanBeEmptyText text={name} />
+    </td>
+    <td>
+      <CanBeEmptyText text={symbol} />
+    </td>
+    <td>{decimals}</td>
+  </>
+);
 
 export default memo(ERC20Item);
