@@ -1,50 +1,51 @@
-import { FC, memo, useContext } from "react";
+import { FC, memo } from "react";
 import DecoratedAddressLink from "../execution/components/DecoratedAddressLink";
 import BlockLink from "../components/BlockLink";
 import TimestampAge from "../components/TimestampAge";
 import CanBeEmptyText from "../components/CanBeEmptyText";
-import { RuntimeContext } from "../useRuntime";
-import {
-  ContractMatch,
-  useBlockData,
-  useERC721Metadata,
-} from "../useErigonHooks";
+import { ChecksummedAddress } from "../types";
+import { ResultMapper } from "../ots2/useUIHooks";
 
 type ERC721ItemProps = {
-  m: ContractMatch;
+  blockNumber: number;
+  timestamp: number;
+  address: ChecksummedAddress;
+  name: string;
+  symbol: string;
 };
 
-const ERC721Item: FC<ERC721ItemProps> = ({ m: { address, blockNumber } }) => {
-  const { provider } = useContext(RuntimeContext);
-  const erc721Meta = useERC721Metadata(provider, address, blockNumber);
-  const block = useBlockData(provider, blockNumber.toString());
+export const mapper: ResultMapper<ERC721ItemProps> = (m, blocksSummary) => ({
+  blockNumber: m.blockNumber,
+  timestamp: blocksSummary.get(m.blockNumber)!.timestamp,
+  address: m.address,
+  name: m.name,
+  symbol: m.symbol,
+});
 
-  return (
-    <tr>
-      <td>
-        <DecoratedAddressLink address={address} plain />
-      </td>
-      <td>
-        <BlockLink blockTag={blockNumber} />
-      </td>
-      <td>{block && <TimestampAge timestamp={block.timestamp} />}</td>
-      {erc721Meta ? (
-        <>
-          <td>
-            <CanBeEmptyText text={erc721Meta.name} />
-          </td>
-          <td>
-            <CanBeEmptyText text={erc721Meta.symbol} />
-          </td>
-        </>
-      ) : (
-        <>
-          <td></td>
-          <td></td>
-        </>
-      )}
-    </tr>
-  );
-};
+const ERC721Item: FC<ERC721ItemProps> = ({
+  blockNumber,
+  timestamp,
+  address,
+  name,
+  symbol,
+}) => (
+  <>
+    <td>
+      <DecoratedAddressLink address={address} plain />
+    </td>
+    <td>
+      <BlockLink blockTag={blockNumber} />
+    </td>
+    <td>
+      <TimestampAge timestamp={timestamp} />
+    </td>
+    <td>
+      <CanBeEmptyText text={name} />
+    </td>
+    <td>
+      <CanBeEmptyText text={symbol} />
+    </td>
+  </>
+);
 
 export default memo(ERC721Item);
