@@ -4,7 +4,7 @@ import {
   BlockWithTransactions,
   BlockTag,
 } from "@ethersproject/abstract-provider";
-import { JsonRpcProvider, TransactionResponse } from "@ethersproject/providers";
+import { JsonRpcProvider } from "@ethersproject/providers";
 import { getAddress } from "@ethersproject/address";
 import { Contract } from "@ethersproject/contracts";
 import { defaultAbiCoder } from "@ethersproject/abi";
@@ -23,8 +23,6 @@ import {
   TokenMeta,
 } from "./types";
 import erc20 from "./erc20.json";
-import erc721md from "./erc721metadata.json";
-import { rawToProcessed } from "./search/search";
 
 const TRANSFER_TOPIC =
   "0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef";
@@ -614,7 +612,9 @@ export const useAddressBalance = (
  * arguments.
  */
 export const providerFetcher =
-  (provider: JsonRpcProvider | undefined): Fetcher<any | undefined, any[]> =>
+  (
+    provider: JsonRpcProvider | undefined
+  ): Fetcher<any | undefined, [string, ...any]> =>
   async (key) => {
     if (provider === undefined) {
       return undefined;
@@ -699,449 +699,6 @@ export const useTokenMetadata = (
     fetcher
   );
   if (error) {
-    return undefined;
-  }
-  return data;
-};
-
-// TODO: fix json names on erigon return
-export type ContractMatch = {
-  blockNumber: number;
-  address: string;
-};
-
-export type TransactionMatch = {
-  hash: string;
-};
-
-const useGenericContractList = (
-  provider: JsonRpcProvider | undefined,
-  rpcMethod: string,
-  pageNumber: number,
-  pageSize: number
-): ContractMatch[] | undefined => {
-  const fetcher = providerFetcher(provider);
-  const { data, error } = useSWRImmutable(
-    [rpcMethod, pageSize, pageNumber - 1],
-    fetcher
-  );
-  if (error) {
-    return undefined;
-  }
-
-  if (data === undefined) {
-    return undefined;
-  }
-  const converted = (data as any[]).map((m) => {
-    const c: ContractMatch = {
-      blockNumber: BigNumber.from(m.Block).toNumber(),
-      address: m.Address,
-    };
-    return c;
-  });
-  return converted;
-};
-
-const useGenericContractCount = (
-  provider: JsonRpcProvider | undefined,
-  rpcMethod: string
-): number | undefined => {
-  const fetcher = providerFetcher(provider);
-  const { data, error } = useSWRImmutable([rpcMethod], fetcher);
-  if (error) {
-    return undefined;
-  }
-  return data as number | undefined;
-};
-
-export const useContractsCount = (
-  provider: JsonRpcProvider | undefined
-): number | undefined => {
-  return useGenericContractCount(provider, "ots_getAllContractsCount");
-};
-
-export const useContractsList = (
-  provider: JsonRpcProvider | undefined,
-  pageNumber: number,
-  pageSize: number
-): ContractMatch[] | undefined => {
-  return useGenericContractList(
-    provider,
-    "ots_getAllContractsPage",
-    pageNumber,
-    pageSize
-  );
-};
-
-export const useERC20Count = (
-  provider: JsonRpcProvider | undefined
-): number | undefined => {
-  return useGenericContractCount(provider, "ots_getERC20Count");
-};
-
-export const useERC20List = (
-  provider: JsonRpcProvider | undefined,
-  pageNumber: number,
-  pageSize: number
-): ContractMatch[] | undefined => {
-  return useGenericContractList(
-    provider,
-    "ots_getERC20Page",
-    pageNumber,
-    pageSize
-  );
-};
-
-export const useERC721Count = (
-  provider: JsonRpcProvider | undefined
-): number | undefined => {
-  return useGenericContractCount(provider, "ots_getERC721Count");
-};
-
-export const useERC721List = (
-  provider: JsonRpcProvider | undefined,
-  pageNumber: number,
-  pageSize: number
-): ContractMatch[] | undefined => {
-  return useGenericContractList(
-    provider,
-    "ots_getERC721Page",
-    pageNumber,
-    pageSize
-  );
-};
-
-export const useERC1155Count = (
-  provider: JsonRpcProvider | undefined
-): number | undefined => {
-  return useGenericContractCount(provider, "ots_getERC1155Count");
-};
-
-export const useERC1155List = (
-  provider: JsonRpcProvider | undefined,
-  pageNumber: number,
-  pageSize: number
-): ContractMatch[] | undefined => {
-  return useGenericContractList(
-    provider,
-    "ots_getERC1155Page",
-    pageNumber,
-    pageSize
-  );
-};
-
-export const useERC1167Count = (
-  provider: JsonRpcProvider | undefined
-): number | undefined => {
-  return useGenericContractCount(provider, "ots_getERC1167Count");
-};
-
-export const useERC1167List = (
-  provider: JsonRpcProvider | undefined,
-  pageNumber: number,
-  pageSize: number
-): ContractMatch[] | undefined => {
-  return useGenericContractList(
-    provider,
-    "ots_getERC1167Page",
-    pageNumber,
-    pageSize
-  );
-};
-
-const useGenericTransactionCount = (
-  provider: JsonRpcProvider | undefined,
-  rpcMethod: string,
-  address: ChecksummedAddress
-): number | undefined => {
-  const fetcher = providerFetcher(provider);
-  const { data, error } = useSWRImmutable([rpcMethod, address], fetcher);
-  if (error) {
-    return undefined;
-  }
-  return data as number | undefined;
-};
-
-const useGenericTransactionList = (
-  provider: JsonRpcProvider | undefined,
-  rpcMethod: string,
-  address: ChecksummedAddress,
-  pageNumber: number,
-  pageSize: number
-): TransactionMatch[] | undefined => {
-  const fetcher = providerFetcher(provider);
-  const { data, error } = useSWRImmutable(
-    [rpcMethod, address, pageNumber - 1, pageSize],
-    fetcher
-  );
-  if (error) {
-    return undefined;
-  }
-
-  if (data === undefined) {
-    return undefined;
-  }
-  const converted = (data as any[]).map((m) => {
-    const t: TransactionMatch = {
-      hash: m.hash,
-    };
-    return t;
-  });
-  return converted;
-};
-
-export const useERC20TransferCount = (
-  provider: JsonRpcProvider | undefined,
-  address: ChecksummedAddress
-): number | undefined => {
-  return useGenericTransactionCount(
-    provider,
-    "ots_getERC20TransferCount",
-    address
-  );
-};
-
-export const useERC20TransferList = (
-  provider: JsonRpcProvider | undefined,
-  address: ChecksummedAddress,
-  pageNumber: number,
-  pageSize: number
-): TransactionMatch[] | undefined => {
-  return useGenericTransactionList(
-    provider,
-    "ots_getERC20TransferPage",
-    address,
-    pageNumber,
-    pageSize
-  );
-};
-
-// TODO: remove temporary prototype
-export const useTransactionsWithReceipts = (
-  provider: JsonRpcProvider | undefined,
-  hash: string[] | undefined
-): ProcessedTransaction[] | undefined => {
-  const fetcher = providerFetcher(provider);
-  const { data, error } = useSWRImmutable(
-    hash === undefined ? null : ["ots_getTransactionsWithReceipts", hash],
-    fetcher
-  );
-  if (error) {
-    return undefined;
-  }
-
-  if (!provider || data === undefined) {
-    return undefined;
-  }
-  const converted = rawToProcessed(provider, data);
-  return converted.txs;
-};
-
-const ERC721MD_PROTOTYPE = new Contract(AddressZero, erc721md);
-
-const erc721MetadataFetcher =
-  (
-    provider: JsonRpcProvider | undefined
-  ): Fetcher<
-    ERC721Metadata | null,
-    ["erc721meta", ChecksummedAddress, number]
-  > =>
-  async ([_, address, blockNumber]) => {
-    if (provider === undefined) {
-      return null;
-    }
-
-    const contract = ERC721MD_PROTOTYPE.connect(provider).attach(address);
-    try {
-      const [name, symbol] = await Promise.allSettled([
-        contract.name({ blockTag: blockNumber + 1 }),
-        contract.symbol({ blockTag: blockNumber + 1 }),
-      ]);
-
-      return {
-        name: name.status === "fulfilled" ? (name.value as string) : "",
-        symbol: symbol.status === "fulfilled" ? (symbol.value as string) : "",
-      };
-    } catch (err) {
-      // Ignore on purpose; this indicates the probe failed and the address
-      // is not a token
-      return null;
-    }
-  };
-
-export type ERC721Metadata = {
-  name: string;
-  symbol: string;
-};
-
-export const useERC721Metadata = (
-  provider: JsonRpcProvider | undefined,
-  address: ChecksummedAddress | undefined,
-  blockNumber: number | undefined
-): ERC721Metadata | undefined | null => {
-  const fetcher = erc721MetadataFetcher(provider);
-  const { data, error } = useSWRImmutable(
-    provider !== undefined && address !== undefined && blockNumber !== undefined
-      ? ["erc721meta", address, blockNumber]
-      : null,
-    fetcher
-  );
-  if (error) {
-    return undefined;
-  }
-  return data;
-};
-
-export const useERC1167Impl = (
-  provider: JsonRpcProvider | undefined,
-  address: ChecksummedAddress | undefined
-): ChecksummedAddress | undefined | null => {
-  const fetcher = providerFetcher(provider);
-  const { data, error } = useSWRImmutable(
-    ["ots_getERC1167Impl", address],
-    fetcher
-  );
-  if (error) {
-    return undefined;
-  }
-  return data;
-};
-
-const erc20MetadataFetcher =
-  (
-    provider: JsonRpcProvider | undefined
-  ): Fetcher<ERC20Metadata | null, ["erc20meta", ChecksummedAddress, number]> =>
-  async ([_, address, blockNumber]) => {
-    if (provider === undefined) {
-      return null;
-    }
-
-    const contract = ERC20_PROTOTYPE.connect(provider).attach(address);
-    try {
-      const [name, symbol, decimals] = await Promise.allSettled([
-        contract.name({ blockTag: blockNumber + 1 }),
-        contract.symbol({ blockTag: blockNumber + 1 }),
-        contract.decimals({ blockTag: blockNumber + 1 }),
-      ]);
-
-      return {
-        name: name.status === "fulfilled" ? (name.value as string) : "",
-        symbol: symbol.status === "fulfilled" ? (symbol.value as string) : "",
-        decimals:
-          decimals.status === "fulfilled"
-            ? (decimals.value as number)
-            : undefined,
-      };
-    } catch (err) {
-      // Ignore on purpose; this indicates the probe failed and the address
-      // is not a token
-      return null;
-    }
-  };
-
-export type ERC20Metadata = {
-  name: string;
-  symbol: string;
-  decimals: number | undefined;
-};
-
-export const useERC20Metadata = (
-  provider: JsonRpcProvider | undefined,
-  address: ChecksummedAddress | undefined,
-  blockNumber: number | undefined
-): ERC20Metadata | undefined | null => {
-  const fetcher = erc20MetadataFetcher(provider);
-  const { data, error } = useSWRImmutable(
-    provider !== undefined && address !== undefined && blockNumber !== undefined
-      ? ["erc20meta", address, blockNumber]
-      : null,
-    fetcher
-  );
-  if (error) {
-    return undefined;
-  }
-  return data;
-};
-
-export const useERC20Holdings = (
-  provider: JsonRpcProvider | undefined,
-  address: ChecksummedAddress
-): ChecksummedAddress[] | undefined => {
-  const fetcher = providerFetcher(provider);
-  const { data, error } = useSWR(["ots_getERC20Holdings", address], fetcher);
-  const converted = useMemo(() => {
-    if (error) {
-      return undefined;
-    }
-
-    if (data === undefined || data === null) {
-      return undefined;
-    }
-    return (data as any[]).map((m) => m.address);
-  }, [data, error]);
-
-  return converted;
-};
-
-const erc20BalanceFetcher =
-  (
-    provider: JsonRpcProvider | undefined
-  ): Fetcher<
-    BigNumber | null,
-    ["erc20balance", ChecksummedAddress, ChecksummedAddress]
-  > =>
-  async ([_, address, tokenAddress]) => {
-    if (provider === undefined) {
-      return null;
-    }
-
-    const contract = ERC20_PROTOTYPE.connect(provider).attach(tokenAddress);
-    return contract.balanceOf(address);
-  };
-
-export const useTokenBalance = (
-  provider: JsonRpcProvider | undefined,
-  address: ChecksummedAddress | undefined,
-  tokenAddress: ChecksummedAddress | undefined
-): BigNumber | null | undefined => {
-  const fetcher = erc20BalanceFetcher(provider);
-  const { data, error } = useSWR(
-    ["erc20balance", address, tokenAddress],
-    fetcher
-  );
-  if (error) {
-    return undefined;
-  }
-
-  if (data === undefined || data === null) {
-    return undefined;
-  }
-  return data;
-};
-
-export type AddressAttributes = {
-  erc20?: boolean;
-  erc165?: boolean;
-  erc721?: boolean;
-  erc1155?: boolean;
-  erc1167?: boolean;
-  erc1167Logic?: boolean;
-};
-
-export const useAddressAttributes = (
-  provider: JsonRpcProvider | undefined,
-  address: ChecksummedAddress
-): AddressAttributes | undefined => {
-  const fetcher = providerFetcher(provider);
-  const { data, error } = useSWR(
-    ["ots_getAddressAttributes", address],
-    fetcher
-  );
-  if (error) {
-    return undefined;
-  }
-
-  if (data === undefined || data === null) {
     return undefined;
   }
   return data;
